Allow presentation in custom windows
This PR allows consumers to pass custom `UIWindow` instances in which to display the tooltips.

The use case I have is showing a custom window behind the tooltip to automatically dismiss the tooltip if the user taps anywhere else on the screen, and showing the tooltip within this window allows for only handling touches that don't overlap with the tooltip. If this functionality is useful, I can also open a PR to add support for it.